Traffic light

+++ Tasks/src/main/java/com/muzhesky/student/Mark.java
@@ -1,5 +1,5 @@
 package com.muzhesky.student;
 
 public interface Mark {
-    public int getValue();
+    public double getValue();
 }
